requirements.txt
streamlit
langchain[llms]
ics
python-dotenv
openai